pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "claude-memory"
version = "1.0.0"
description = "Persistent, human-curated memory for Claude Code"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "claude-memory contributors" },
]
keywords = ["claude", "memory", "ai", "developer-tools"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "pyyaml>=6.0",
]

[project.scripts]
claude-memory = "claude_memory.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
claude_memory = ["skeleton/**/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
